% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_one.R
\name{levels}
\alias{levels}
\alias{level_one}
\alias{level_two}
\alias{level_three}
\title{Level one test}
\usage{
level_one(n = 5)

level_two(n = 5)

level_three(n = 5)
}
\arguments{
\item{n}{Number of questions}
}
\description{
Level one test
}
