% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_level.R
\name{run_level}
\alias{run_level}
\title{Produce questions in console}
\usage{
run_level(n, functions, arguments)
}
\arguments{
\item{n}{Number of questions.}

\item{functions}{Vector of function names.}

\item{arguments}{Vector of argument names.}
}
\description{
Must be used in a level_*() function. Cannot be used on its own.
}
